Minimal lovable product.
The code is now capable of running cogwheel as a pipeline (that is,
without the need to write a script for each analysis as is suggested
by the cogwheel documentation).

Some of the more advanced features of the full code are also
implemented, including distance marginalisation and prior choices,
however setting prior bounds is not yet fully supported.